requirements.txt
# Document Parsing
pdfplumber>=0.10.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.1.0

# Language Detection
langdetect>=1.0.9

# LLM - Google Gemini
google-genai>=1.0.0

# Configuration
PyYAML>=6.0.1

# Utilities
tqdm>=4.66.0
